# Improve performance of genomicLocationShouldAnnotate