--- a/core/src/main/java/org/mskcc/cbio/oncokb/cache/CacheFetcher.java
+++ b/core/src/main/java/org/mskcc/cbio/oncokb/cache/CacheFetcher.java
@@ -122,11 +122,7 @@ public class CacheFetcher {
                     continue;
                 }
                 String chromosome = ensemblGene.getChromosome();
-
-                if (!chromosomeEnsemblGenesMap.containsKey(chromosome)) {
-                    chromosomeEnsemblGenesMap.put(chromosome, new HashSet<>());
-                }
-
+                chromosomeEnsemblGenesMap.putIfAbsent(chromosome, new HashSet<>());
                 chromosomeEnsemblGenesMap.get(chromosome).add(ensemblGene);
             }
         }