--- a/core/src/main/java/org/mskcc/cbio/oncokb/util/MainUtils.java
+++ b/core/src/main/java/org/mskcc/cbio/oncokb/util/MainUtils.java
@@ -1,5 +1,6 @@
 package org.mskcc.cbio.oncokb.util;
 
+import org.apache.commons.lang3.StringUtils;
 import org.cbioportal.genome_nexus.model.GenomicLocation;
 import org.mskcc.cbio.oncokb.apiModels.*;
 import org.mskcc.cbio.oncokb.model.*;
@@ -863,13 +864,22 @@ public class MainUtils {
     }
 
     public static int findDigitEndIndex(String str, int startIndex) {
+        if (startIndex < 0) {
+            return -1;
+        }
         int i = startIndex;
         while (i < str.length() && Character.isDigit(str.charAt(i))) {
             i++;
         }
-        return i;
+        return i == startIndex ? -1 : i;
     }
 
+    /**
+     * We don't intend to check if the HGVSg is valid. If you need a method that validates and parses,
+     * then use the NotationConverter.parseHGVSg() method.
+     * @param hgvsg 
+     * @return null if cannot parse, otherwise the GenomicLocation
+     */
     public static GenomicLocation parseChromosomeAndRangeFromHGVSg(String hgvsg) {
         if (hgvsg == null) {
             return null;
@@ -879,26 +889,34 @@ public class MainUtils {
 
         // Step 1: Split by ":g."
         String[] parts = hgvsg.split(":g\\.");
+        if (parts.length != 2 || parts[0].isEmpty() || parts[1].isEmpty()) {
+            return null;
+        }
         location.setChromosome(parts[0]);  // Chromosome is the part before ":g."
         String coordinates = parts[1];  // This is the part after "g."
 
         // Step 2: Handle the coordinates part
         int underscoreIndex = coordinates.indexOf('_');
-
-        if (underscoreIndex != -1) {
-            // If there is an underscore, we have both start and end values
-            start = Integer.parseInt(coordinates.substring(0, underscoreIndex));
-            // Find where the digits after the underscore end (before any letters)
-            int endIndex = findDigitEndIndex(coordinates, underscoreIndex + 1);
-            end = Integer.parseInt(coordinates.substring(underscoreIndex + 1, endIndex));
-        } else {
-            // No underscore means start = end
-            int endIndex = findDigitEndIndex(coordinates, 0);
-            start = Integer.parseInt(coordinates.substring(0, endIndex));
-            end = start;
+        try {
+            if (underscoreIndex != -1) {
+                // If there is an underscore, we have both start and end values
+                start = Integer.parseInt(coordinates.substring(0, underscoreIndex));
+                // Find where the digits after the underscore end (before any letters)
+                int endIndex = findDigitEndIndex(coordinates, underscoreIndex + 1);
+                end = Integer.parseInt(coordinates.substring(underscoreIndex + 1, endIndex));
+            } else {
+                // No underscore means start = end
+                int endIndex = findDigitEndIndex(coordinates, 0);
+                start = Integer.parseInt(coordinates.substring(0, endIndex));
+                end = start;
+            }
+            location.setStart(start);
+            location.setEnd(end);
+        } catch (NumberFormatException exception) {
+            return null;
+        } catch (IndexOutOfBoundsException exception) {
+            return null;
         }
-        location.setStart(start);
-        location.setEnd(end);
         return location;
     }
 }