Related to https://github.com/oncokb/oncokb-pipeline/issues/553

### Before Improvements

#### Performance Benchmark
- WES - 13100 variants 31s (422 v/s); 1127 variants were within oncokb coverage (8%) 
- WGS - 1447800 variants: 1362s (1062v/s); 43993 variants were within oncokb coverage (3%) 
![image](https://github.com/user-attachments/assets/df868858-790c-49b2-828d-9a9701e829ce)

### After Improvements

#### Changes:
1. Create a `Map<String, Set<EnsemblGene>>`, where the key is the chromosome. This allows us to only search the set of EnsemblGenes for one chromosome and allows for short circuiting. There is no reason to check the range if the chromosome doesn’t match the query. 
2. For each reference genome, create two maps. This will reduce the size of Set<EnsemblGene> by ½. 
3. Eliminate the NototationConverter.hgvsgToGenomicLocation() call in oncokb core. It seems that this conversion is only used so that we can grab the chromosome and the start+end to check if there is a range intersect. This can be done without the use of regex.

#### Performance Benchmark
- WES – 13100 variants 25s (524v/s) 
- WGS – 1447800 variants: 1019s (1420v/s) 

![image](https://github.com/user-attachments/assets/dca66da4-8c83-4954-aeb1-b55931c4d3e7)