@calvinlu3 2024-10-23
@zhx828 With in-memory cache, it takes takes 25s to annotate 13100 variants (524v/s). With redis, it takes 26.89s to annotate 13100 variants (487v/s). About a 7% decrease in performance